fix(#104): creates unique folder for each test project in Test Bed
